Uploaded Internal Assessment Documentation
This is for proof of completion of the product as well as documenting the different phases of the SDLC. Criteria A was gathering the requirements for the system, Criteria B was creating the design and creating flowcharts, Criteria C was the development of the system, Criteria D was video proof of the working code, and Criteria E was the evaluation of the whole process. The appendix contains screenshots of all the code as well as communications with my client and there are product screenshots available.